SearchKeywords as interface #1053
